+++ src/Vizy.php
@@ -55,7 +55,7 @@ class Vizy extends Plugin
     // =========================================================================
 
     public bool $hasCpSettings = true;
-    public string $schemaVersion = '1.0.1';
+    public string $schemaVersion = '1.0.2';
 
 
     // Public Methods